requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
pypdf
gradio